Improve the way scalar coercion happens
It is legitimate for a scalar type to want to coerce nils into non-nil values
and vice-versa, in normal (non-error) cases. This means:
- calling `coerce_result` from a method called `non_null_result` is misleading
at best, since it is legitimate for `coerce_result` to return nil
- the short-circuiting that `BaseResolution#result` does on nils is incorrect
for scalars since it doesn't give the coercion an opportunity to handle that
case

Simply override `result` instead of `non_null_result` for `ScalarResolution`.
Still short-circuit when the raw value is an `ExecutionError` but otherwise
trust the coercion proc. Update the string and id scalars to pass through nils
since otherwise they turned them into `""` which is probably undesirable.