Allow specifying return interfaces for mutations
It is simple to define informal interfaces in mutation results
but it's nicer to have this supported in the type system.

It is currently possible to specify that a mutation payload
type supports certain interfaces but that requires implementing
the whole Payload type from scratch, with this change it's as simple
as declaring interface support for other types, while still
allowing the same simple definition for non-interface fields.