Make LazyMethodMap thread-safe using Concurrent::Map as backing storage
As identified in #628, the previous LazyMethodMap implementation was not
thread-safe, which could cause issues when using graphql and a ruby
threaded webserver (runtime errors and potentially initializing the
same location more than once).

This affected both MRI and JRuby.

To fix this, this commit switches LazyMethopMap's backing storage from a
regular ruby Hash into a Concurrent::Map from the concurrent-ruby gem
(adding this gem as a dependency), thus making this class thread-safe.

Fixes #628