Add UnionType#resolve_type for more ganular control in unions
The purpose of this feature is to allow a union type to resolve to
more specific types when handling a value. This is useful when working
with objects that may fit into multiple types. When executing the query
will call resolve_type(value, field_ctx) on either Unions or Interfaces.

Union type's call can be definable via `resolve_type ->(value, ctx) { }`.
Interface type's call returns `ctx.query.resolve_type(value)` which is the
default implementation for resolving both types.