ActionCableChannelOperationsHandler
The current action cable implementation creates a new “channel” for
each relay requestSubscription call. Internally, action cable tracks
each channel which is identified by the channel name “GraphqlChannel”.

When a message is received, js action cable looks for all the local js
channels subscribed to “GraphqlChannel” and delivers the server payload
to each one.

Unfortunately, this means that for each payload received, it will be
sent to the observer of every single requestSubscription.

This change addresses that by asking the client to forward a
channel_operations_id to the server. This is similar to how
subscriptions-transport-ws works. The server is responsible for sending
the channel operations id back to client. The client uses this id
internally to map a specific observer to an observer.