Implement `.hash` on Nodes, enable Ruby 2.5
AST nodes were implementing `.eql?` without implementing `.hash`.
`.eql?` is supposed to return true if an object is exactly the same, or
if its contents are exactly the same.  This is used to resolve hash
collisions.  Any object whose `.hash` is the same, the hash
implementation will call `.eql?` on them to determine if those to
objects are *exactly* the same (if so, then one is discarded as it's a
dupe).  In this case, we were implementing `.eql?` without implementing
`.hash`, so nodes that are semantically equivalent would return true for
`.eql?` but would not produce the same `.hash` value.  Two objects that
return true for `eql?` must also hash to the same value, so this patch
implements the `.hash` method.

Implementing the `.hash` method causes the test to fail the same way on
Ruby 2.4 as it does on Ruby 2.5, so I updated the test to fix the
assertion.