Properly validate interfaces without selections
Previously this was only checking object types. This means a query
containing a field which returns an interface type is valid without any
selections. This is wrong according to the spec and the reference JS
implementation though.

This updates the validation rule to use `fields?` instead of `object?`.

The error message has also been updated to reflect it's not just object
types and a helpful "Did you mean" part was borrowed from graphql-js.