Revert extending interface methods onto Objects
This is effectively a revert of #1635.

When an Object Type would implement an Interface, the interface's
`DefinitionMethods` were being extended.

However this causes unexpected ancestor chain issues when dealing with
common methods (such as `to_graphql`) which are common to both
interfaces and objects.

This simply removes this feature. Objects no longer extend the
definition methods of the interfaces they implement.