Add inaccessible_field feature
Right now accessibility is done at query analysis time. The only
provided hook is `inaccessible_fields` where you can either allow the
query to continue or raise an error.

If you raise an error, query execution does not occur. This prevents any
partial responses from happening.

If `nil` is returned and the query continues, all information from
`GraphQL::Authorization::Analyzer` is lost.

The idea behind this is to provide a 2nd option where the authorization
enforcement is done during field execution (but before actually
resolving) to allow for partial responses.

Ideally this would be opt-in at the schema level by providing a
`inaccessible_field` (note the difference, no `s` at the end, name TBD)
method which would be called for *each* field which is not accessible.
An `ExecutionError` could be raised to prevent the field was being
resolved.

Things to figure out:

1. how to opt-in to this behaviour. The optional schema method doesn't
actually work due to the legacy schema and delegation.2
2. this only works for class-based fields. Does this need to be
supported for the legacy ones as well? Is there even a way to do this
nicely?
3. is checking the `irep_node` a stable enough way to do this?