Improve support for extensions with options
Currently its only possible to specify one extension with options when
using the `extensions` kwarg in a field and it has to be the *last*
extension specified.

This improves handling of `extension` and allows for a mix of extension
classes and extensions with options which no longer have to come last.