Add error handling to correctly report coercion errors in mutations
We are currently pinned to a version of graphql-ruby prior to a bug fix
for the handling of coercion errors in mutation processing, until the
transition to Rails 4, and a subsequent upgrade of Ruby, is complete.

This commit applies that fix to the older version that we are using. The
fix is the one to the lib/grapqhl/query/variables.rb file here:

https://github.com/rmosolgo/graphql-ruby/pull/1602/commits/d20e28041f56c4f7f14aba6a123cdefa84779151

We will create a gem from this branch that will be used by
api.epublishing.com.

Re: https://epublishing.atlassian.net/browse/CMS-701