Fix argument deserialization
As outlined in https://github.com/rmosolgo/graphql-ruby/issues/2611
there appears to be a bug in argument deserialization if the argument
contains a number. This bug was confirmed by @rmosolgo [here](https://github.com/rmosolgo/graphql-ruby/issues/2611#issuecomment-558790828)

This commit contains a patch for the bug. The culprit was
`GraphQL::Schema::Member::BuildType#underscore`. Upon investigation,
this method used gsub` to include underscores but just split the string
based on letters. This was confirmed by adding a test for the
`#underscore` method in `spec/graphql/schema/member/build_type_spec.rb`
which was not initially present.

After identifying the issue, the following two `gsub` calls were chained
to the `underscore` method:

```ruby
.gsub(/([A-Za-z]+)([0-9]+)/,'\1_\2')  # some1 -> some_1
.gsub(/([0-9]+)([A-Za-z]+)/,'\1_\2')  # 1thing -> 1_thing
```

Having added the tests that confirm the bug has been fixed, a number of
other tests failed due to the test code using the incorrectly
deserialized values. This commit also patches these tests.