Put a hard limit on number of conflicts to report
This change limits the number of conflicts we will report to users.

Currently, conflict search is inherently `O(n^2)`, since we use nested
loops to compare like against like. This exhaustive search is of limited
use if there are many conflicts, and represents a vulnerability if the
input is malicious.

To address this, the number of conflicts we will consider is capped,
using the existing `ValidationContext.max_errors` configuration.