Lint lexer.rb and parser.rb after they are generated
The generated files are not properly indented, which generates a lot of warnings
when loaded by client apps. (Ref. https://github.com/rmosolgo/graphql-ruby/issues/1256)

It may (should?) be possible to correct the indentations by hand-adjusting the grammar
files but rubocop does the job admirably well, and as the grammar files are rarely touched
(only a couple of commits per year have modified the lexer/parser in the last few years),
it seems like an acceptable option to me as it gets rid of the warnings just as well.

We only make layout autocorrections (-x) to err on the side
of caution, although it does miss adding the frozen_string_literal comment.