Pass and merge subscription's event and query ctxs
When a subscription trigger is executed, there's an param to pass context but it's lost because it's not forwarded in the `Subscription::Event`. 
Now when `Subscription#execute_update` is executed, we can now merge query's context with event's context. 
This allows to pass an id of the sender of a query and compare it with the receiver of a subscription in the `#update` method, for example to prevent the sender of the subscription to receive it's own updates by skipping it when the ids match.

Example:

````ruby
module Subscriptions
  class ExampleSubscription < ::Types::Base::Subscription
    payload_type ::Types::Models::ExampleSubscriptionPayload

    argument :model_id, ID, required: true, loads: ::Types::Models::MyModel, can_can_action: :read

    def subscribe(model:)
      super
    end

    def update(journey:)
      # skip update when the sender is the same as the receiver of the update
      return :no_update if self.context[:client_id].present? && self.context[:sender_client_id].present? && self.context[:client_id] == self.context[:sender_client_id]

      super
    end
  end
end
````