Make GraphQL parser cache opt-in
The cache has nothing to do with bootsnap, so we don't need to piggyback
off of it at all. This will prevent bootsnap cache to be polluted by the
graphql gem, and allow more apps to leverage GQL caching. The parser
cache is also now opt-in, which will make this behaviour more expected.