Silence Ruby warnings when redefining field accessors
A relatively common pattern is to redefine a field accessor, e.g.:

```ruby
field :billing_address,
  -> { MailingAddressInput },
  required: true,
  description: "The customer's billing address."

def billing_address
  Address.new(self[:billing_address].to_h)
end
```

Unfortunately this emits a Ruby warning.

To avoid it, we can either define these methods in a module
that is included in the class, or use the self alias trick.

I chose to do the later as it is simpler.