Suppress warning for objectId arguments
Using an argument named `objectId` currently emits a warning [1] because
it redefines the `#object_id` method. This should technically be safe to
do because Ruby doesn't use that method internally, but it still warns
because third party code might.

`objectId` is a valid name for an argument, and a useful one too. This
commit suppresses the warning by temporarily setting `$VERBOSE` to
`nil` while defining the method. It's unlikely any code is relying on
the `#object_id` method, but if so it's already broken—this commit just
silences the warning.

[1]: https://github.com/ruby/ruby/blob/1001ea9606fd9bbbce70deaab0e6a31c5af1a20c/vm_method.c#L1066-L1070