Fix order-dependency bug in required validator
When using the required validator, if the first option is a set and it
is matched, the validation will always be valid, even if other options
in the mutually exclusive group would match. This is due to the
presence of a `break` statement, which will cause a premature exit if a
set option is matched.