Improve default InvalidNullError handling
`InvalidNullError` was modeled as a `RuntimeTypeError` but also
implemented `to_h` to act like an `ExecutionError`. It was a strange
hybrid which made it harder for applications to change how the errors
were handled.

It also caused limitations such as not being easy to add `extensions` or
`path` to the error in the response.

This changes `InvalidNullError` to inherit from the base
`GraphQL::Error` instead of `RuntimeTypeError`. Instead of adding the
instance of `InvalidNullError` directly to `context.errors`, the default
implementation is changed to create a new `ExecutionError` instead.

This provides a simpler middle ground where applications can still
choose to handle `InvalidNullError` exceptions differently (such as
reporting them to their exception handler app) but keep the proper spec
compliant field error response.