Update Int and Float coercion logic to be spec compliant
The spec says that scalars should return query/field errors when input
or result coercion values don't conform to the spec.

This updates the lgoic to return `GraphQL::CoercionError` (execution
errors) instead of raised `RuntimeError`s which would not result in a
client error by default.

The logic and error messages are were taken from graphql-js.

Note that this is likely a breaking change for schema developers. For
clients, it's an improvement since they might get some sort of "internal
srever error" otherwise.