Ensure `data` entry exists for execution errors
According to the spec, the `data` entry should always exist once
execution starts. Only if an error is raised _before_ execution starts
should the `data` key be omitted.

The only test that needed updating was when an `ExecutionError` was
raised during the `execute_query` trace hook. Despite this running
before `super`, this still happens "during execution" and should have
the `data` key present.

Ref: https://spec.graphql.org/draft/#sec-Data